work on TIFF, mainly dedicated to handle large files - fix bug with offsets being bounded or converted to NA_integer. offsets where stored as uint32_t in hpp/cpp and placed in Rcpp::IntegerVector and then converted with as.integer in R which makes them undesirably bound to [-(2^31-1) : -2^31-1]). In hpp/cpp we now use std::size_t to access files at desired position and store offset in Rcpp::NumericVector. In R, we remove as.integer conversion. - current IFD offset serves as reference for next computed offset which are supposed to be greater than this reference (except for 0x00000000 last one) and next offset (for IFD or extra content) are computed accordingly - improvement in offset extraction. -- dry code for hpp/cpp offset extraction -- Now offsets are faster to extract, notably when initial number of object is not known thanks to vector preallocation (before push_back was used) -- drop RcppProgress dependency. RcppProgress was used to monitor the extraction of offset within hpp/cpp. We now call internal IFC:::setPB from hpp/cpp function so all progress bar are now homogeneized within IFC - speed improvement during file export (ExportToXIF, subsetXIF, mergeXIF, XIFtoTIFF). Though it still remains long. - order of images files export has changed. Previously, extra content was written before IFD. Now, IFD is written first and then content that does not fit within the 4 bytes slot for value is written just after the offset pointing to next IFD
file I/O from hpp/cpp
- fix bug that prevented file from being released when an error raised during reading at hpp/cpp level
- use std::size_t instead of uint32_t for offset position (in scan.hpp and decomp.hpp, in addition to tiff.hpp)

data export
- dry code, data_to_DAF now makes use of toBIN_features and toBIN_images intenal function
- order of images files export has changed. Previously, extra content was written before IFD. Now, IFD is written first and then content that does not fit within the 4 bytes slot for value is written just after the offset pointing to next IFD

typo in doc (missing param in toBIN functions, UNDIFINED in getIFD)

+++ R/checksumDAF.R
@@ -108,7 +108,7 @@ checksumDAF <- function(fileName, endianness = .Platform$endian, ...) {
     ##### loop over xml SO nodes for id and img/msk offsets
     while((length(obj) != 0) && (i_image < SO_number)) {
       i_image = i_image + 1
-      val = as.integer(xml_attrs(nodes[[i_image]])[1:2])
+      val = as.numeric(xml_attrs(nodes[[i_image]])[1:2])
       if(val[1] %in% obj) {
         if(warn) if(val[1] != obj[1]) { # ensure it is stored in ascending order
           warning("raw object are not stored in expected order")
